Studio - Move StudioWrapper into src/apps/studio and repoint the route shim
Relocates StudioWrapper.tsx into src/apps/studio and points src/apps/studio/index.tsx at ./StudioWrapper, so the Studio app is now self-contained. Import paths only; no behaviour change.
